+++ client/tws_foundation_client/lib/src/services/bases/carriers_service_base.dart
@@ -1,0 +1,15 @@
+import 'package:csm_client/csm_client.dart';
+import 'package:tws_foundation_client/tws_foundation_client.dart';
+
+
+abstract class CarriersServiceBase extends CSMServiceBase {
+  CarriersServiceBase(
+    super.host,
+    super.servicePath, {
+    super.client,
+  });
+
+  /// Transaction to generate a set view object.
+  Effect<SetViewOut<Carrier>> view(SetViewOptions<Carrier> options, String auth);
+
+}