+++ client/tws_foundation_client/lib/src/sets/business/sets_business_module.dart
@@ -24,6 +24,7 @@ export 'trailer_type.dart';
 export 'truck.dart';
 export 'truck_common.dart';
 export 'truck_external.dart';
+export 'truck_inventory.dart';
 export 'usdot.dart';
 export 'vehicule_model.dart';
 export 'yard_log.dart';