+++ server/TWS Foundation/TWS Business/Depots/CarriersDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class CarriersDepot : BDepot<TWSBusinessDatabase, Carrier> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="Carrier"/>.
     /// </summary>
+    public CarriersDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+       : base(Databases, Disposer) {
+    }
     public CarriersDepot() : base(new(), null) {
     }
 }