+++ server/TWS Foundation/TWS Business/Depots/LocationsDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class LocationsDepot : BDepot<TWSBusinessDatabase, Location> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="Location"/>.
     /// </summary>
+    public LocationsDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+        : base(Databases, Disposer) {
+    }
     public LocationsDepot() : base(new(), null) {
     }
 }