+++ server/TWS Foundation/TWS Business/Depots/PlatesDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -12,7 +13,9 @@ public class PlatesDepot
     /// <summary>
     ///     Generates a new depot handler for <see cref="Plate"/>.
     /// </summary>
-
+    public PlatesDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+       : base(Databases, Disposer) {
+    }
     public PlatesDepot()
         : base(new(), null) {
     }