+++ server/TWS Foundation/TWS Business/Depots/TrailersExternalsDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class TrailersExternalsDepot : BDepot<TWSBusinessDatabase, TrailerExterna
     /// <summary>
     ///     Generates a new depot handler for <see cref="TrailerExternal"/>.
     /// </summary>
+    public TrailersExternalsDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+       : base(Databases, Disposer) {
+    }
     public TrailersExternalsDepot() : base(new(), null) {
     }
 }